Fix drawers for Ellipse, Rectangle, Text and Line

+++ src/main/java/cpaintus/controllers/drawers/PokeballDrawerStrategy.java
@@ -25,7 +25,7 @@ public class PokeballDrawerStrategy implements IDrawerStrategy {
 		double width = shape.getWidth();
 		double height = shape.getHeight();
 		
-		double lineWidth = shape.getLineWidth();
+		int lineWidth = shape.getLineWidth();
 		
 		// Colored top of the ball
 		gc.fillArc(lineWidth/2,
